with the egg pole design implementation started.

+++ lib/models/bill_item.dart
@@ -2,10 +2,10 @@
 class BillItem {
   final String productId;
   final int qty;
-  final double price;
+  final int price;
 
-  /// Eggs consumed per unit of this product — copied from [Product.eggsPerUnit]
-  /// at bill-creation time so old bills remain self-contained.
+  /// Eggs consumed per unit of this product — captured at bill-creation time
+  /// so old bills remain self-contained.
   final int eggsPerUnit;
 
   const BillItem({
@@ -15,22 +15,22 @@ class BillItem {
     this.eggsPerUnit = 0,
   });
 
-  double get lineTotal => qty * price;
+  int get lineTotal => qty * price;
 
   /// Total eggs consumed by this line item.
   int get totalEggs => qty * eggsPerUnit;
 
   factory BillItem.fromJson(Map<String, dynamic> json) => BillItem(
-        productId: json['productId'] as String,
-        qty: json['qty'] as int,
-        price: (json['price'] as num).toDouble(),
-        eggsPerUnit: json['eggsPerUnit'] as int? ?? 0,
-      );
+    productId: json['productId'] as String,
+    qty: json['qty'] as int,
+    price: (json['price'] as num).toInt(),
+    eggsPerUnit: json['eggsPerUnit'] as int? ?? 0,
+  );
 
   Map<String, dynamic> toJson() => {
-        'productId': productId,
-        'qty': qty,
-        'price': price,
-        'eggsPerUnit': eggsPerUnit,
-      };
+    'productId': productId,
+    'qty': qty,
+    'price': price,
+    'eggsPerUnit': eggsPerUnit,
+  };
 }
